Even Liber Abbaci

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -9,7 +9,7 @@ int main(void)
 {
 	long int a, n = 4000000;
 	long int f, b = 0, c = 1;
-	long long int summ = 0;
+	long int summ = 0;
 
 	for (a = 1; a < n; a++)
 	{
@@ -21,6 +21,6 @@ int main(void)
 			summ = summ + f;
 		}
 	}
-	printf("%lld\n", summ);
+	printf("%ld\n", summ);
 	return (0);
 }
